Split graph related code out of nav.dhcpstats.common
