+++ python/nav/models/manage.py
@@ -57,7 +57,7 @@ from nav.models.fields import DateTimeInfinityField, VarcharField, PointField
 from nav.models.fields import CIDRField
 import nav.models.event
 from nav.oids import get_enterprise_id
-import nav.dhcpstats.common
+import nav.dhcpstats.graph
 
 
 _logger = logging.getLogger(__name__)
@@ -1568,7 +1568,7 @@ class Prefix(models.Model):
         range/pool/subnet intersects this prefix.
         """
         prefix = IPy.IP(self.net_address)
-        return nav.dhcpstats.common.fetch_graph_urls_for_prefixes([prefix])
+        return nav.dhcpstats.graph.fetch_graph_urls_for_prefixes([prefix])
 
 
 class Vlan(models.Model):
@@ -1676,7 +1676,7 @@ class Vlan(models.Model):
         range/pool/subnet intersects this vlan.
         """
         prefixes = [IPy.IP(prefix.net_address) for prefix in self.prefixes.all()]
-        return nav.dhcpstats.common.fetch_graph_urls_for_prefixes(prefixes)
+        return nav.dhcpstats.graph.fetch_graph_urls_for_prefixes(prefixes)
 
 
 class NetType(models.Model):
